# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aerokod Test API",
        "version": "1.0"
    },
    "paths": {
        "/api": {},
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Получение списка пользователей",
                "operationId": "ad744e0f103fb22b172e0b9864845006",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Jessica Smith"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "example@example.com"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-04-19T08:12:42.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Создание пользователя",
                "operationId": "6a21bb5dabd0ec5b55a59ad96a6e7fb2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jessica Smith"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "example@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jessica Smith"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "example@example.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-04-19T08:12:42.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{user}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Получение одного пользователя",
                "operationId": "50a2f5c49a8c8d0f5196d3ae6a0af6bb",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jessica Smith"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "example@example.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-04-19T08:12:42.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Обновление пользователя",
                "operationId": "01226f407236066849ff71a55686a271",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jessica Smith"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jessica Smith"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "example@example.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-04-19T08:12:42.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Удаление пользователя",
                "operationId": "e4db18fe4be7b008ac6741596be49f3a",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Пользователь удален"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}